# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Teaching Planning Assistant API",
    "description": "Autonomous educational content generation engine for Chilean curriculum",
    "version": "2.0.0"
  },
  "paths": {
...
    "/admin/apikeys/": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List Api Keys",
        "description": "List all API keys (Admin only)",
        "operationId": "list_api_keys_admin_apikeys__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyResponse"
                  },
                  "type": "array",
                  "title": "Response List Api Keys Admin Apikeys  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key (Admin only). The API key is only shown once!",
        "operationId": "create_api_key_admin_apikeys__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAPIKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAPIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/admin/apikeys/{client_id}": {
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Deactivate Api Key",
        "description": "Deactivate an API key (Admin only)",
        "operationId": "deactivate_api_key_admin_apikeys__client_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/apikeys/{client_id}/rotate": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Rotate Client Api Key",
        "description": "Rotate (regenerate) an API key (Admin only)",
        "operationId": "rotate_client_api_key_admin_apikeys__client_id__rotate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAPIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/batch/generate": {
      "post": {
        "tags": [
          "Batch Operations"
        ],
        "summary": "Create Batch Job Endpoint",
        "description": "Submit a batch content generation job.\n\nCreates a batch job that will process multiple content generation\nrequests asynchronously. Progress and completion are reported via\nwebhook if configured.\n\nAuthentication: API Key with batch:generate permission",
        "operationId": "create_batch_job_endpoint_api_v1_batch_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/batch/{job_id}": {
      "get": {
        "tags": [
          "Batch Operations"
        ],
        "summary": "Get Batch Status",
        "description": "Get status of a batch job.\n\nReturns current progress and optionally results if completed.\n\nAuthentication: API Key",
        "operationId": "get_batch_status_api_v1_batch__job_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "include_results",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Results"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batch/{job_id}/cancel": {
      "post": {
        "tags": [
          "Batch Operations"
        ],
        "summary": "Cancel Batch Job Endpoint",
        "description": "Cancel a pending or processing batch job.\n\nAuthentication: API Key",
        "operationId": "cancel_batch_job_endpoint_api_v1_batch__job_id__cancel_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batch/": {
      "get": {
        "tags": [
          "Batch Operations"
        ],
        "summary": "List Batch Jobs",
        "description": "List batch jobs for the current client.\n\nAuthentication: API Key",
        "operationId": "list_batch_jobs_api_v1_batch__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchJobInfo"
                  },
                  "title": "Response List Batch Jobs Api V1 Batch  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "rate_limit": {
            "type": "integer",
            "title": "Rate Limit"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_used": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_name",
          "permissions",
          "rate_limit",
          "is_active",
          "created_at",
          "last_used",
          "webhook_url"
        ],
        "title": "APIKeyResponse"
      },
...
      "BatchContentType": {
        "type": "string",
        "enum": [
          "quiz",
          "activity",
          "exam",
          "reinforcement",
          "lesson"
        ],
        "title": "BatchContentType",
        "description": "Types of content that can be batch generated"
      },
      "BatchGenerateRequest": {
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/BatchContentType",
            "description": "Type of content to generate"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items",
            "description": "Items to generate (each with item_id and content-specific fields)"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url",
            "description": "URL for progress/completion notifications"
          },
          "webhook_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Secret",
            "description": "Secret for webhook signature"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "description": "low, normal, high",
            "default": "normal"
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "items"
        ],
        "title": "BatchGenerateRequest",
        "description": "Request model for batch generation"
      },
      "BatchGenerateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BatchJobInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "BatchGenerateResponse",
        "description": "Response model for batch generation request"
      },
      "BatchJobInfo": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "completed_items": {
            "type": "integer",
            "title": "Completed Items"
          },
          "failed_items": {
            "type": "integer",
            "title": "Failed Items"
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "content_type",
          "total_items",
          "completed_items",
          "failed_items",
          "progress_percent",
          "created_at",
          "started_at",
          "completed_at"
        ],
        "title": "BatchJobInfo",
        "description": "Information about a batch job"
      },
      "BatchStatusResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/BatchJobInfo"
          },
          "results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "job"
        ],
        "title": "BatchStatusResponse",
        "description": "Response for batch job status"
      },
...
      "CreateAPIKeyRequest": {
        "properties": {
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "default": [
              "*"
            ]
          },
          "rate_limit": {
            "type": "integer",
            "title": "Rate Limit",
            "default": 100
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          }
        },
        "type": "object",
        "required": [
          "client_name"
        ],
        "title": "CreateAPIKeyRequest"
      },
      "CreateAPIKeyResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "rate_limit": {
            "type": "integer",
            "title": "Rate Limit"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_used": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_name",
          "permissions",
          "rate_limit",
          "is_active",
          "created_at",
          "last_used",
          "webhook_url",
          "api_key"
        ],
        "title": "CreateAPIKeyResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}